[S1.1] Move resolveDkgCli + resolveCliPackageDir to dkg-core
S1 step 1 of execution-plan.md §3.S1 (issue #386 — Hermes setup parity).
Behavior-unchanged extraction; both helpers move from
adapter-openclaw to @origintrail-official/dkg-core so that adapter-hermes
can reuse them in S2. Dependency direction is cli → adapters → core,
so dkg-core (not dkg-cli) is the only valid shared home — confirmed
with team-lead during S1 prep.

- Added: packages/core/src/resolve-cli-package-dir.ts
- Added: packages/core/src/resolve-dkg-cli.ts
- Added: exports for both via packages/core/src/index.ts
- adapter-openclaw/src/setup.ts now thin-re-exports resolveCliPackageDir
  from dkg-core (preserves the adapter's public surface)
- adapter-openclaw/src/resolve-dkg-cli.ts now thin-re-exports
  resolveDkgCli from dkg-core (preserves the in-tree import path used
  by setup.ts and setup-start-daemon.test.ts)
- adapter-openclaw/test/resolve-dkg-cli.test.ts mock target updated
  from '../src/setup.js' to the dkg-core resolve-cli-package-dir.js
  module path (vitest mocks need a real cross-module boundary, so the
  two helpers live in two separate core modules)

Test results vs s1-baseline.md (post-extraction, all targeted packages):
- @origintrail-official/dkg-core: 33 files, 550 tests — green
- @origintrail-official/dkg-adapter-openclaw: 22 files, 974 tests —
  same 2 pre-existing failures as baseline (plugin-id reconcile pending,
  writeDkgConfig autoUpdate edge); zero new failures
- @origintrail-official/dkg-adapter-hermes: 1 file, 60 tests — green
- @origintrail-official/dkg-node-ui: 20 files — green

Note on @origintrail-official/dkg (CLI) tests: blocked on a baseline
infrastructure issue (Hardhat global setup fails to spin up the local
chain on port 9548), which is independent of this extraction. Flagging
to team-lead in parallel.
