+++ packages/core/src/index.ts
@@ -26,6 +26,8 @@ export {
   blueGreenSlotReady,
 } from './blue-green.js';
 export { requestFaucetFunding, type FaucetResult } from './faucet.js';
+export { resolveCliPackageDir } from './resolve-cli-package-dir.js';
+export { resolveDkgCli, type ResolvedDkgCli } from './resolve-dkg-cli.js';
 export {
   assertSafeIri,
   isSafeIri,
